+++ src/errors/NotFoundError.ts
@@ -1,8 +1,8 @@
 import { InternalError } from "./InternalError";
 
 export class NotFoundError extends InternalError {
-	constructor() {
-		super(404, "Not found");
+	constructor(res: Response) {
+		super(res, 404, "Not found");
 		this.name = "NotFoundError";
 	}
 }