+++ src/routes/index.test.ts
@@ -1,7 +1,0 @@
-import { routes } from "./index";
-
-describe("router", () => {
-	test("doesn't contain invalid routes", () => {
-		expect(Object.keys(routes)).not.toContain("");
-	});
-});